Initial upload.
Contains the following:

Defines the data structure that will be used for assembler instructions, code
segments, loops and functions.
Defines the resources the 8085 system has and their involvement and regulation
of usage in the delay function.
Defines the interface the InstructionSet need to provide to be functional (this
is device dependent).
Proposes a brute force method for generating delay functions.

Does not contain:

A model for presenting the instruction set data- which instructions are
available, what they do how long they last etc.
The dynamic algorithm for solving the problem of using n-instructions that last
exactly pre-specified time.
A genetic evolution approach (maybe it is not required at all brute force could
work, requested additional calculations).
- This will require an evaluation function
- Combining model
- Mutating model
- Generating random code-segments