Finally after a lot of duck-taping on the code we can generate some delay functions.
The part for separating the available instructions from the whole set is done.
The code needs to be re-factored especially he following:
--the instruction set, all the functions for fetching return instructions and such are moved to the CodeBuilder and hardcoded for now, a nicer solution is required
--in the buildLoop function which is very important there is some serious code repetition,
deal with that ASAP
--another case of code repetition is in the buildFunction and buildLoop, deal with that as well

Instructions that should not be used when generating code such as JMP, CALL, RET and such should be somehow tagged in the instruction Metadata text file, and that information forwarded to all the relevant places.

Instruction class should have a time variable. Fix that and all the time functions in all Executable extendees (children classes).

No testing has been done with memory or stack instructions.

Further notice: It's decided that the function should try to release some registers by saving and restoring their contents to another register, memory or stack, in order to create a working space for its logic.